requirements.txt
azure-functions
azure-storage-blob
azure-servicebus
azure-cosmos
python-dateutil
